ubc: Remove native implementations The initial motivation for introducing native implementation for the UBC checks, was to act as a stepping stone on a fully native implementation of sha1cd.
That is no longer being pursued, for two main reasons:
- The non-trivial complexity that comes with a full-on goasm version with
collision detection, which would make it harder to maintain/debug the project.
- The recently added SIMD implementations made the project reasonably fast
for the core supported architectures (arm64 and amd64).

It is important to also note that the performance gained with the native UBC
versions were minor, certainly not worth the complexity they added to the project.
