--- ubc/ubc_noasm.go
@@ -1,14 +1,0 @@
-//go:build (!amd64 && !arm64) || noasm || !gc
-// +build !amd64,!arm64 noasm !gc
-
-package ubc
-
-// CalculateDvMask takes as input an expanded message block and
-// verifies the unavoidable bitconditions for all listed DVs. It returns
-// a dvmask where each bit belonging to a DV is set if all unavoidable
-// bitconditions for that DV have been met.
-//
-//go:nosplit
-func CalculateDvMask(W [80]uint32) uint32 {
-	return CalculateDvMaskGeneric(W)
-}
